+++ data-query/src/test/resources/gov/va/api/health/dataquery/service/controller/diagnosticreport/v1/datamart-diagnostic-report.json
